# Slotted array search and insert optimizations